requirements.txt
# pip install

bcrypt
flask
Flask-PyMongo
Flask-Login
gunicorn

